requirements.txt
altair==5.5.0
annotated-types==0.7.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
filelock==3.19.1
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
hf-xet==1.2.0
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
llvmlite==0.45.1
MarkupSafe==2.1.5
more-itertools==10.8.0
mpmath==1.3.0
narwhals==2.9.0
networkx==3.5
numba==0.62.1
numpy==2.3.4
openai-whisper==20250625
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.0
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.10.23
reportlab==4.4.4
requests==2.32.5
rpds-py==0.28.0
safetensors==0.6.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sounddevice==0.5.3
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
